refactor: Invitation 도메인을 클린 아키텍쳐 패키지로 변경 #37
- 패키지 버저닝 추가
